Updated 'defaults' related commands and docs

+++ scripts/defaults.sh
@@ -6,20 +6,39 @@ source _utils.sh
 e_pending "Creating defaults"
 # ------------------------------------------------------------------------------
 
-defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
-defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
-defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
-defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
-defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
-defaults write com.apple.dock autohide -boolean true
-killall Dock
+get_consent "Create Dock spacers"
+  if has_consent; then
+    e_pending "Creating Dock spacers"
+    defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
+    defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
+    defaults write com.apple.dock persistent-apps -array-add '{"tile-type"="spacer-tile";}'
+    killall Dock
+  fi
+fi
 
-defaults write com.apple.finder AppleShowAllFiles -boolean true
-killall Finder
+get_consent "Autohide Dock"
+  if has_consent; then
+    e_pending "Autohiding Dock"
+    defaults write com.apple.dock autohide -boolean true
+    killall Dock
+  fi
+fi
+
+get_consent "Display hidden Finder files/folders"
+  if has_consent; then
+    e_pending "Displaying hidden Finder files/folders"
+    defaults write com.apple.finder AppleShowAllFiles -boolean true
+    killall Finder
+  fi
+fi
 
 if ! has_path "Projects"; then
-  mkdir -p ~/Projects
-  test_path "Projects"
+  get_consent "Create ~/Projects folder"
+  if has_consent; then
+    e_pending "Creating ~/Projects folder"
+    mkdir -p ~/Projects
+    test_path "Projects"
+  fi
 fi
 
 if ! has_command "xcode-select"; then
